Security hardening: fix 20 findings, mandate C++ memory-safety skills, add sanitizer/tidy/fuzz gates (#39)
* Install curl, lzma, clang and clang-tidy dev packages in the agent environment


* Vendor cpp-coding-standards, cpp-sanitizers and memory-safety-patterns skills

Pinned to upstream commits recorded in each LICENSE.txt (all MIT).


* Mandate the C++ memory-safety skills through an always-on rule


* Add hardening flags, EMULOCKE_SANITIZE builds and pin FetchContent dependencies to commits

Project targets compile with -Wall -Wextra -Wformat=2, stack protector and
_GLIBCXX_ASSERTIONS; Linux Release adds _FORTIFY_SOURCE=3 and full RELRO.
EMULOCKE_SANITIZE=address|thread applies ASan+UBSan or TSan to the whole
build and wires runtime options plus suppression files into ctest.


* Copy assets through one target instead of three racing POST_BUILD copies

Parallel builds intermittently failed with "Error copying directory" when
emulocke, emulocke-sprite-check and emulocke-game-art-check copied assets
into the same directory at once.


* Bound BPS, UPS and xdelta patch application against hostile patches

BPS: variable-length integers stop at ten bytes and reject overflow, every
command checks length and offsets in subtraction form, the target size is
capped by kMaxPatchedRom, and TargetCopy may only read already-written
output. The mGBA IPS/UPS path applies the same output cap, and its VFile is
owned by a unique_ptr. xdelta no longer retries with Adler32 verification
disabled and bounds its output buffer growth.

patch_bounds_check exercises each rejection path; the pre-fix code aborted
with std::bad_alloc on an oversized target and accepted checksum-failed
xdelta output.


* Check the full Gen4 partition extent before reading a save

readGen4Save only required 0x1000 bytes past storageStart, then read 18 boxes
of Pokémon plus box names (about 74 KiB) from a possibly truncated file.
Under ASan the old reader segfaults on a save cut just past that check; the
layout now declares its true extent and Gen5 names its minimum size instead
of relying on an offset that happened to cover every read.


* Release mGBA config on GbaSession teardown and quiet xdelta3 alignment checks

LeakSanitizer flagged 144 allocations from mCoreInitConfig that deinit never
freed. The ROM-load failure path now also closes the VFile it still owns.
xdelta3 performs deliberate unaligned loads, so its target opts out of the
UBSan alignment check under EMULOCKE_SANITIZE.


* Reject PK4/PK5 blobs larger than the decrypt buffer

parsePk45 handed decryptPk45 a span claiming raw.size() bytes over a fixed
236-byte std::array, so any longer input was a stack-buffer-overflow (ASan:
WRITE of size 300). The buffer is now passed honestly and oversized input is
rejected; party stats are read only when the blob actually contains them.


* Cap every whole-file read and take size_t lengths when writing

readWholeFile requires a caller-named ceiling (ROM 512 MiB, save 16 MiB,
patch 64 MiB, image 32 MiB), rejects a failed tellg instead of turning -1
into a huge allocation, and fails on short reads. writeWholeFile takes
std::size_t so patched ROMs no longer truncate through uint32_t.


* Limit PNG decode size and dimensions and cap HTTPS downloads

decodePngRgba rejects buffers above kMaxImageFile and any side above 4096
(also enforced inside stb_image through STBI_MAX_DIMENSIONS), and owns the
stb allocation through unique_ptr. httpGetPng stops at 32 MiB, follows at
most five redirects, allows https only, and verifies the host name. The
new media check covers giant IHDR, truncated header, oversized buffer and
a real round trip.


* Parse prefs.ini with from_chars and clamp every numeric value

atoi has undefined behaviour on out-of-range digits and silently turned
garbage into 0. Parsing now lives in Prefs::parse over any istream so it
can be tested; malformed values keep their defaults and window geometry,
scale, volume and speed-up are clamped to their valid ranges.


* Own SDL through RAII: SdlRuntime, unique_ptr window and renderer, atomic audio state

SDL_Quit ran inside Application::shutdown while ScreenTexture, Input and
the remaining unique_ptr members were still alive and destroyed SDL objects
afterwards. SdlRuntime is now the first member so it is torn down last, Host
holds its window and renderer in unique_ptr with custom deleters (the
renderer-failure path no longer strands a window and destroy() only shuts
down ImGui it actually initialised), and AudioOutput's cross-thread flags
are std::atomic. The new lifetime check runs under SDL's dummy drivers and
passes under ASan and LeakSanitizer.


* Allowlist slug bytes, bound NDS RAM reads, reject .. in melonDS local paths and reread the battery only when it changes


* Add libFuzzer harnesses for patch and save parsers with seed corpora and a smoke script


* Pin Actions to commit SHAs, gate uploads to push, add sanitize, clang-tidy and fuzz-smoke CI jobs


* Document the 2026-09-14 security audit and point the memory-safety rule at the fuzz workflow


* Silence xdelta3 zero-offset null-pointer UB under clang UBSan


* Rename Cursor rule to securitystandards.mdc so it no longer shares a skill name

Rules stay project-generic and point at the vendored skills; the downloaded skill directories keep their upstream names.


* Cap xdelta3 allocations so a huge VCDIFF app header cannot abort ASan

fuzz_xdelta died on a mutated patch whose app-header VLI made xd3_alloc request ~505 TiB. Decode through the stream API with an allocator that returns null above kMaxRomFile, and keep the CI crash unit in the corpus.


* Silence clang-tidy WaE hits in calc after merging the Calculator board

SessionSearch now compares std::search to end without an unqualified auto iterator. Stat stages in Gen3Atk are converted to int explicitly so signed-char-misuse does not treat int8_t stages as character bytes.


* Silence xdelta3 memcpy-from-null UBSan on empty decode windows

fuzz_xdelta aborted when xd3_process_memory called memcpy with a null next_out. Exclude nonnull-attribute on the xdelta3 target (same policy as its other known UB) and keep the CI crash unit in the corpus.


* Turn off UBSan for xdelta3 so the fuzzer stops dying on library UB

ASan stays on. The 60s smoke kept tripping a different xdelta3 UBSan check each run (null memcpy, then a DJW destroy function-pointer type mismatch). Those are library bugs, not holes in applyXdeltaPatch.


---------

+++ src/calc/Gen3Atk.cpp
@@ -43,7 +43,7 @@ int advAttack(const AdvCtx& ctx) {
     if (ctx.physical && (a.ability == kAbHustle || (a.ability == kAbGuts && a.status))) {
         at = at * 3 / 2;
     }
-    const int stage = ctx.physical ? a.atkStage : a.spaStage;
+    const int stage = ctx.physical ? static_cast<int>(a.atkStage) : static_cast<int>(a.spaStage);
     if (stage > 0 || (!ctx.crit && stage < 0)) {
         at = boostedStat(at, stage);
     }
@@ -62,7 +62,7 @@ int advDefense(const AdvCtx& ctx) {
     if (ctx.physical && d.ability == kAbMarvelScale && d.status) {
         df = df * 3 / 2;
     }
-    const int stage = ctx.physical ? d.defStage : d.spdStage;
+    const int stage = ctx.physical ? static_cast<int>(d.defStage) : static_cast<int>(d.spdStage);
     if (stage < 0 || (!ctx.crit && stage > 0)) {
         df = boostedStat(df, stage);
     }
